third_party/blink/web_tests/external/wpt/payment-request/idlharness.https.window-expected.txt

This is a testharness.js-based test.
Found 1 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] PaymentRequest interface object length
  assert_equals: wrong value for PaymentRequest.length expected 2 but got 1
Harness: the test ran to completion.
